refactor TpRestApiLogInterceptor
